Make cBioPortal mutations dataset from Synapse assets

The NF-OSI workflow produces a single merged maf file that represents a filtered subset of the mafs, containing only the (non-germline) data that can be released for cBioPortal. However, this data file by itself is not immediately loadable into an instance of a cBioPortal server and needs to be packaged with other files, such as this example of a public mutations dataset. This is a wrapper that goes through several steps needed to create said bundle of cBioPortal files more conveniently.

Usage

syncBP_maf(
  merged_maf,
  samplesheet,
  ref_map,
  ref_view,
  name,
  cancer_study_identifier,
  citation = NULL,
  pmid = NULL,
  short_name = NULL,
  publish_dir = cancer_study_identifier,
  verbose = TRUE
)

Arguments

merged_maf

Synapse id of merged maf file for public release.

samplesheet

Synapse id or local path to samplesheet with release info.

ref_map

YAML file specifying the mapping of (NF) clinical metadata to cBioPortal model. See details.

ref_view

A view that contains clinical data for the release files.

name

Name of the cancer study, e.g. something following convention is "Malignant Peripheral Nerve Sheath Tumor (NF-OSI, 2022)".

cancer_study_identifier

Study identifier, convention is {tumorType}_{institution}_{year}, so for example "mpnst_nfosi_2022".

citation

(Optional) A relevant citation, e.g. "TCGA, Nature 2012".

pmid

(Optional) One or more relevant pubmed ids (comma separated without whitespace); if used, citation cannot be NULL.

short_name

(Optional) Short name for the study.

publish_dir

Where to output the set of files. Defaults to (creating if necessary) a folder with same name as cancer_study_identifier.

verbose

Whether to provide informative messages throughout.

Details

A simple sanity check that this is the version of the maf release file that we want, based on the samplesheet. For example, version 1 of the samplesheet will generate a version 1 of the merged maf, but if there is a later correction that retracts a sample (is_releasable=FALSE), that step of the workflow to generate merged maf will be rerun (or it should be), so we'll want to make sure the latest versions of these files are used. The latest versions of the samplesheets tied to each release are currently stored in syn38793855. (Note: Please file issue to update this doc if this changes.)

Make the clinical data files. For NF, the clinical metadata are annotations on the files/surfaced in a view and are pretty basic. In the future, it would be preferable to store clinical metadata in a real normalized table. For now, this clinical data is pulled in from the view. To map NF clinical variables to the cBioPortal dictionary as recommended, this step requires a ref_map, which is a YAML file.

Make meta files. Meta files are needed for describing the study, mutations data file, clinical data files.