Annotate somatic or germline variants output

Currently, vcf are output first and mafs appear after a subsequent workflow run, so depending on when map_sample_output_sarek is run to create a sample_io file, there will be just vcf files or both. (In the future, the workflows will likely be joined so both are deposited at the same run.) One can specify to only annotate either the "vcf" or "maf" files and create a manifest for just those files, or use "auto" to detect the file types present in sample_io.

Usage

annotate_called_variants(
  sample_io,
  workflow_ref,
  format = c("auto", "vcf", "maf"),
  template = "bts:ProcessedVariantCallsTemplate",
  schema =
    "https://raw.githubusercontent.com/nf-osi/nf-metadata-dictionary/main/NF.jsonld",
  data_type = c("auto", "SomaticVariants", "GermlineVariants",
    "AnnotatedSomaticVariants", "AnnotatedGermlineVariants"),
  verbose = TRUE,
  dry_run = TRUE
)

Arguments

sample_io

Table mapping input to outputs, which reference output .vcf.gz or maf files.

workflow_ref

Character vector with names for workflow and values as version-specific links.

format

Variant format, "auto" to handle any "vcf" or "maf" files present automatically, or specify one explicitly. See details.

template

(Optional) URI of template in data model to use, prefixed if needed. Can specify different model/version, but in some cases may not work well.

schema

Path (URL or local) to file from which schema will be read, or schema as list object.

data_type

Variant type, use "auto" to infer from naming scheme and current NF processing SOP, or specify more explicitly.

verbose

Give verbose reports for what's happening.

dry_run

Whether to apply annotations.

Details

maf files use the same template but with different default values, such as a subclass term for dataType. If in the future mafs require a significantly different template, then this should be factored out into a separate annotation function.