Map sample to output from nf-sarek

See https://nf-co.re/sarek. Most processed outputs are nested by sample and variant callers, i.e. *VariantCalling/<TUMOR_vs_NORMAL>/<CALLER>. Other times the data is organized as *VariantCalling/<CALLER>/<TUMOR_vs_NORMAL>. This walks through the output destination (URI of *VariantCalling) with similar intention to map_sample_output_rnaseq, but for Sarek outputs.

Usage

map_sample_output_sarek(syn_out, sample_level = 2)

Arguments

syn_out

Syn id of syn output destination with files of interest.

sample_level

If caller is organized by sample, use 2 (default), if samples organized by caller, use 3. See details.

Value

A data.table with cols caller caller_path caller_syn output_name output_id sample workflow

Details

Note: And additional step post-Sarek will create MAFs in the output subdirectory DeepVariant. If this is run after the MAF creation step, this will return file indexes with those .maf files.